Support CloudFoundry out of the box
This commit enables client and server to work out of the box on
CloudFoundry.
The client derives the registration info (url and metadata) from the
VCAP environment variables.
The server computes a unique instanceId and the correct headers for the
cf router based on the metadata provided by instances. Also suppots the
CloudFoundry discovery client ootb.

closes #434