Fix snapshot build instructions
The previous code snippet contained code which would go in the `distributionManagement` section of a POM, meant for where to publish snapshots. It is however intended for users to use in their POMs to access published snapshot versions of the Spring Boot Admin artifacts. This change makes that more clear and gives the correct snippet to copy-and-paste into one's POM.